完凸状況を見れる機能追加。

+++ ManageActions.py
@@ -18,63 +18,32 @@ class ManageActions:
 
     def check_and_action(self, req):
 
-#        # NOTE:
-#        # <@478542546537283594> hoge という形でreq.contentに入ってくるため、hoge以前の文字列を削除
-#        message = re.sub('\<.+\> ', '', req.content)
-#
-#        # NOTE:
-#        # サーバに所属するメンバーを管理するテーブルの作成、および、メンバーの追加
-#        if re.match("^init$", message):
-#            cm = Common.Common()
-#            ws = cm.get_gsfile('clan_members')
-#            df = cm.create_gsdf(ws)
-#
-#            # シートclan_membersの内容を初期化（クリア）
-#            col_lastnum = len(df.columns)
-#            row_lastnum = len(df.index)
-#
-#            cell_list = ws.range('A2:'+cm.toAlpha(col_lastnum)+str(row_lastnum + 1))
-#            for cell in cell_list:
-#                cell.value = ''
-#
-#            ws.update_cells(cell_list)
-#            time.sleep(5)
-#            df_col = df.columns
-#            del df
-#
-#            # 新たにサーバ上にいるメンバーを格納したDataFrameを作成
-#            df = pd.DataFrame(index=[], columns=df_col)
-#            for m in req.server.members:
-#                s = pd.Series([m.id, m.name], index=df_col)
-#                df = df.append(s, ignore_index=True)
-#
-#            col_lastnum = len(df.columns)
-#            row_lastnum = len(df.index)
-#
-#            cell_list = ws.range('A2:'+cm.toAlpha(col_lastnum)+str(row_lastnum + 1))
-#            for cell in cell_list:
-#                val = df.iloc[cell.row - 2][cell.col - 1]
-#                cell.value = val
-#
-#            ws.update_cells(cell_list)
+        message = re.sub('\<.+\> ', '', req.content)
+        # サーバに所属するメンバーを管理するテーブルの作成、および、メンバーの追加
+        if re.match("^init$", message):
 
+            args = {
+                "host"    : os.getenv("DB_HOST", ""),
+                "user"    : os.getenv("DB_USER", ""),
+                "passwd"  : os.getenv("DB_PASS", ""),
+                "db"      : os.getenv("DB_NAME", ""),
+                "charset" : os.getenv("DB_CHARSET", ""),
+            }
+            
+            with closing(MySQLdb.connect(**args)) as conn:
+                cur = conn.cursor()
+                cur.execute("TRUNCATE TABLE clan_members")
 
-        args = {
-            "host"    : os.getenv("DB_HOST", ""),
-            "user"    : os.getenv("DB_USER", ""),
-            "passwd"  : os.getenv("DB_PASS", ""),
-            "db"      : os.getenv("DB_NAME", ""),
-            "charset" : os.getenv("DB_CHARSET", ""),
-        }
-        
-        with closing(MySQLdb.connect(**args)) as conn:
-            cur = conn.cursor()
-            cur.execute("TRUNCATE TABLE clan_members")
+                for m in req.server.members:
+                    is_member = 0
+                    for r in m.roles:
+                        # TODO:環境変数からとってくるようにする
+                        if r.id == '631094092050202626':
+                            is_member = 1
 
-            for m in req.server.members:
-                cur.execute("INSERT INTO clan_members (user_id, user_name) VALUES (%s, %s)", (m.id, m.name))
+                    cur.execute("INSERT INTO clan_members (user_id, user_name, is_member) VALUES (%s, %s, %s)", (m.id, m.name, is_member))
 
-            conn.commit()
+                conn.commit()
